requirements.txt
django-extensions
ipython
psycopg2-binary==2.9.1
django-admin-decorators==0.1
